Bug fixes to tools.py and rate.py
When reading results for the rate constant finished is now set to True assuming the file is created?

added return statement which was lost before and fixed a bug concerning indexing when reading trajectory files with read_wham_input and the customized version. I have added this before but it did not come trough